requirements.txt
Django~=4.1.1
pytz==2018.5
djangorestframework~=3.13.1
requests~=2.28.1